Moves contact form into separate component folder
